refactor(frontend): extract EnvironmentMultiSelect and DatabaseMode to flatten MembersPage import cycle
