+++ frontend/src/react/pages/settings/types.ts
@@ -1,0 +1,1 @@
+export type DatabaseMode = "ALL" | "EXPRESSION" | "SELECT";
